bulk data generation function

Construct Pseudo bulk samples by random sampled number of cells per subject, but based on the actual numbers.

generateBulk(
eset,
ct.varname,
sample,
disease = NULL,
ct.sub,
prop_mat = NULL,
nbulk = 50,
samplewithRep = F,
low_s = 0.3,
upp_s = 0.7
)

Arguments

eset
ExpressionSet object for single cells

ct.varname
variable name for 'cell types'

sample
variable name for subject/samples

disease
indicate the health condition of subjects

ct.sub
a subset of cell types that are selected to construct pseudo bulk samples. If NULL, then all cell types are used.

prop_mat
manually input the cell-type proportion for pseudo bulk samples

nbulk
number of pseudo bulk samples to be constructed

samplewithRep
logical, randomly sample single cells with replacement. Default is F.

low_s
lower supprot a for uniform distribution U[a,b]

upp_s
upper supprot b for uniform distribution U[a,b]

Value

pseudo bulk samples ExpressionSet, and actual cell-type proportions